split main into 2; removed old main
